refactor: Перенёс CategoryDtoRequest, CategoryDtoResponse
