+++ ewm-main-service/src/main/java/ru/practicum/category/dto/CategoryDtoRequest.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto.category;
+package ru.practicum.category.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
