+++ ewm-main-service/src/main/java/ru/practicum/category/service/AdminCategoryService.java
@@ -1,7 +1,7 @@
 package ru.practicum.category.service;
 
-import ru.practicum.dto.category.CategoryDtoRequest;
-import ru.practicum.dto.category.CategoryDtoResponse;
+import ru.practicum.category.dto.CategoryDtoRequest;
+import ru.practicum.category.dto.CategoryDtoResponse;
 
 /**
  * @author Nikolay Radzivon
