+++ ewm-main-service/src/main/java/ru/practicum/dto/event/EventFullDtoResponse.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.dto.category.CategoryDtoResponse;
+import ru.practicum.category.dto.CategoryDtoResponse;
 import ru.practicum.dto.user.UserShortDto;
 import ru.practicum.json.CustomLocalDateTimeSerializer;
